Latest functional version (now with normalize and denormalize at testing!)
