+++ cnn_despeckling/utils.py
@@ -1,5 +1,10 @@
-import numpy as np
 from scipy import special
+import numpy as np
+from tqdm import tqdm
+from PIL import Image
+import os
+
+from cnn_despeckling.Jarvis import *
 
 # Extract relevant parameters
 L = 1
@@ -13,5 +18,98 @@ def normalize(batch):
 def denormalize(batch):
     # return np.exp((M - m) * np.clip(np.squeeze(im), 0, 1) + m) + 1e-6
     # return np.exp(((M - m) * np.squeeze(im) + m + c))
-    # return np.exp(((M - m) * np.clip(np.squeeze(batch), 0, 1) + m))
-    return np.exp(2 * batch * (M - m) + m + c)
+    # return np.exp(((M - m) * np.clip(np.squeeze(batch), 0, 1) + m)) + 1e-7
+    return np.exp(2 * np.clip(np.squeeze(batch), 0, 1) * (M - m) + m + c) + 1e-7
+    # return np.exp(2 * np.squeeze(batch) * (M - m) + m + c) + 1e-7
+
+def process_patches_with_model(patch_folder, model, device, desc='patches SLA', patch_index=None):
+    processed_patches = []
+
+    # Get all .npy files and sort them
+    patch_files = [f for f in os.listdir(patch_folder) if f.endswith('.npy')]
+    patch_files.sort(key=lambda x: int(os.path.splitext(x)[0].split('_')[-1]))  # Sorting by index
+
+    if patch_index is not None:
+        patch_files = [patch_files[patch_index]]
+
+    for filename in tqdm(patch_files, desc='Processing patches ...'):
+        # Load patch and convert to PyTorch tensor
+        patch_path = os.path.join(patch_folder, filename)
+        patch = np.load(patch_path)
+        patch = normalize(patch)
+
+        patch_tensor = torch.from_numpy(patch).unsqueeze(0)  # Add batch dimension
+
+        # Move tensor to GPU if available
+        patch_tensor = patch_tensor.to(device)
+        # Process with the model
+        with torch.no_grad():
+            model.eval()
+            output_tensor = model(patch_tensor)
+
+        # Convert output tensor to NumPy array and move to CPU
+        output_array = output_tensor.squeeze(0).cpu().numpy()  # Remove batch dimension
+        output_array = denormalize(output_array)
+        processed_patches.append(output_array)
+
+    return processed_patches
+
+def reconstruct_image_from_processed_patches(processed_patches, original_dims, patch_size=256, stride=64, desc='patches SLA'):
+    (original_height, original_width) = original_dims
+    reconstructed = np.zeros((original_height, original_width), dtype=np.float32)
+    counts = np.zeros_like(reconstructed, dtype=np.int32)
+
+    current_x, current_y = 0, 0
+    for patch in tqdm(processed_patches, desc=desc):
+        # Ensure patch is squeezed from any extra dimensions
+        patch = np.squeeze(patch)
+
+        # Place the patch in the reconstructed image
+        reconstructed[current_x:current_x + patch_size, current_y:current_y + patch_size] += patch
+        counts[current_x:current_x + patch_size, current_y:current_y + patch_size] += 1
+
+        # Update coordinates for the next patch
+        current_y += stride
+        if current_y + patch_size > original_width:
+            current_y = 0
+            current_x += stride
+            if current_x + patch_size > original_height:
+                break  # Finished processing all patches
+
+    # Normalize to average overlapping regions
+    reconstructed /= np.maximum(counts, 1)  # Avoid division by zero
+    return np.sqrt(np.abs(reconstructed))
+
+def store_data_and_plot(im, threshold, filename):
+    im = np.clip(im, 0, threshold)
+    im = im / threshold * 255
+    im = Image.fromarray(im.astype(np.uint8)).convert('L')
+    im.save(filename + '.png')
+
+def calculate_enl(image, mask=None):
+    """
+   Calculates the Equivalent Number of Views (ENL) for a SAR image.
+
+    Args:
+    - image: An array of NumPy representing the SAR image.
+    - mask: A NumPy boolean array of the same size as `image`, where True indicates the pixels belonging to the homogeneous region.
+            the pixels belonging to the homogeneous region.
+
+    Returns:
+    - ENL: The calculated ENL value for the specified region.
+    """
+    if mask is None:
+        # If no mask is provided, the entire image is used.
+        region = image.flatten()
+    else:
+        # Selects only the pixels of the homogeneous region.
+        region = image[mask]
+
+    # Calculate the average and variance of the homogeneous region.
+    mean = np.mean(region)
+    variance = np.var(region)
+
+    # Calculate the ENL.
+    enl = (mean ** 2) / variance
+
+    return enl
