+++ config.json
@@ -5,15 +5,15 @@
       "m":-1.429329123112601,
       "MODEL" : "Autoencoder_Wilson_Ver1",
       "ARCHIVE": " Jarvis",
-      "CKPT": "../cnn_despeckling/mis_checkpoints/WilsonVer1_Net_l2_log_loss_leaky_relu_10_20_0.001-v1.ckpt",
+      "CKPT": "../cnn_despeckling/mis_checkpoints/WilsonVer1_Net_mse_leaky_relu_10_30_0.001.ckpt",
       "INPUT_FOLDER": "../data/patches/General_A",
       "FUNCTION": "leaky_relu",
       "REFERENCE_FOLDER": "../data/patches/General_B",
-      "SELECT": "l2_log_loss",
+      "SELECT": "mse",
       "learning_rate": 1e-3,
       "WEIGHT_DECAY": 0.0,
       "batch_size": 10,
-      "epochs": 20,
+      "epochs": 30,
       "ONLYTEST" :false,
       "NUMWORKERS": 16,
       "training_data_percentage": 1
